refactor: update subscription checkout process and remove deprecated files
- Removed the .mcp.json configuration file as it is no longer needed.
- Enhanced the SubscriptionCheckoutPage to validate scheduling period dates before proceeding with checkout.
- Updated the ExpertProfile component to pass scheduling period parameters to the subscription checkout URL.
- Replaced ConsultationPricing component with ExpertPricing to better reflect the context of the expert's services.
- Deleted unused ConsultationPricing and PricingToggle components to streamline the codebase.
- Adjusted subscriptionSearchParamsSchema to include optional scheduling period fields for improved validation.
